requirements.txt
asgiref==3.6.0
Babel==2.12.1
boto3==1.26.114
botocore==1.29.114
dj-database-url==0.5.0
Django==3.2
django-allauth==0.41.0
django-bootstrap-datepicker-plus==5.0.3
django-countries==7.5.1
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.2
django-storages==1.13.2
django-summernote==0.8.20.0
gunicorn==20.1.0
jmespath==1.0.1
oauthlib==3.2.2
phonenumberslite==8.13.10
Pillow==9.5.0
psycopg2==2.9.6
pydantic==1.10.7
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
s3transfer==0.6.0
sqlparse==0.4.3
